fix: critical typo
Manual backend setup docs specify the following:

the correct Bitcoin Core RPC credentials are specified in CORE_RPC
the correct BACKEND is specified in MEMPOOL:
"electrum" if you're using romanz/electrs or cculianu/Fulcrum
"esplora" if you're using mempool/electrs
"none" if you're not using any Electrum Server

However if you set esplora (mempools fork) a lot of these endpoints are incorrectly cut off. This seems to be an oversight, as you'd probably want the opposite to happen.

As a result, manual deployments with esplora as backend as the docs say do not work properly.